main/requirements.txt
# Flask and related extensions
flask
flask_cors
flask_socketio

# Asynchronous networking library
eventlet

# Serial communication library
pyserial